Move User class properties to Abstract User class
