+++ config/doctrine/User/AbstractUser.orm.xml
@@ -6,10 +6,21 @@
 
     <mapped-superclass name="App\Entity\User\AbstractUser">
         <field name="username" type="string" length="255" unique="true"/>
-        <field name="password" type="string" length="255" nullable="true"/>
+        <field name="email" type="string" length="255"/>
+        <field name="externalId" type="string" length="64" unique="true"/>
+
+        <one-to-one field="permissions" target-entity="App\Entity\Permissions\Permissions" orphan-removal="true">
+            <join-column name="permissions_id" referenced-column-name="id" nullable="false"/>
+            <cascade>
+                <cascade-persist/>
+            </cascade>
+        </one-to-one>
+
+        <field name="avatarHash" type="string" length="512" nullable="true"/>
 
         <indexes>
             <index columns="username"/>
+            <index columns="external_id"/>
         </indexes>
     </mapped-superclass>
