+++ config/doctrine/User/User.orm.xml
@@ -4,21 +4,6 @@
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
                           https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
-    <entity name="App\Entity\User\User" table="users" repository-class="App\Repository\UserRepository">
-        <field name="email" type="string" length="255"/>
-        <field name="externalId" type="string" length="64" unique="true"/>
-        <field name="avatarHash" type="string" length="512" nullable="true"/>
-
-        <one-to-one field="permissions" target-entity="App\Entity\Permissions\Permissions" orphan-removal="true">
-            <join-column name="permissions_id" referenced-column-name="id" nullable="false"/>
-            <cascade>
-                <cascade-persist/>
-            </cascade>
-        </one-to-one>
-
-        <indexes>
-            <index columns="external_id"/>
-        </indexes>
-    </entity>
+    <entity name="App\Entity\User\User" table="users" repository-class="App\Repository\UserRepository"/>
 
 </doctrine-mapping>
