+++ src/Entity/AbstractEntity.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace App\Entity;
 
+use App\Entity\User\UserInterface;
 use Ramsey\Uuid\Uuid;
-use Symfony\Component\Security\Core\User\UserInterface;
 
 class AbstractEntity implements EntityInterface
 {
