+++ src/Entity/EntityInterface.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Entity;
 
-use Symfony\Component\Security\Core\User\UserInterface;
+use App\Entity\User\UserInterface;
 
 interface EntityInterface
 {
